+++ src/app/app.ts
@@ -1,15 +1,5 @@
-///<reference path="../../node_modules/grafana-sdk-mocks/app/headers/common.d.ts" />
-
-import kbn from 'app/core/utils/kbn';
-import {
-    loadPluginCss,
-    MetricsPanelCtrl
-} from "app/plugins/sdk";
-import TimeSeries from "app/core/time_series2";
 import { BoomTablePattern } from './BoomTablePattern';
-import * as utils from "./utils";
 
-const plugin_id = "yesoreyeram-boomtable-panel";
 const defaultPattern = new BoomTablePattern({
     bgColors: "green|orange|red",
     bgColors_overrides: "0->green|2->red|1->yellow",
@@ -21,7 +11,7 @@ const defaultPattern = new BoomTablePattern({
     name: "Default Pattern",
     null_color: "darkred",
     null_value: "No data",
-    pattern : "*",
+    pattern: "*",
     row_name: "_series_",
     thresholds: "70,90",
     time_based_thresholds: [],
@@ -29,68 +19,7 @@ const defaultPattern = new BoomTablePattern({
     transform_values_overrides: "0->down|1->up",
     valueName: "avg"
 });
-const config: any = {
-    debug_mode: false,
-    editorTabs: [{
-        name: "Patterns",
-        position: 2,
-        template: "/partials/patterns.html"
-    },
-    {
-        name: "Time based thresholds & Filters",
-        position: 3,
-        template: "/partials/patterns-advanced-options.html"
-    }, {
-        name: "Options",
-        position: 4,
-        template: "/partials/options.html"
-    }],
-    error: undefined,
-    groupedData: undefined,
-    hide_first_column: false,
-    hide_headers: false,
-    list_of_stylesheets: {
-        dark: "plugins/" + plugin_id + "/css/default.dark.css",
-        light: "plugins/" + plugin_id + "/css/default.light.css"
-    },
-    panelDefaults: {
-        activePatternIndex: -1,
-        defaultPattern: defaultPattern,
-        default_title_for_rows: "Metric",
-        nullPointMode: "connected",
-        patterns: [],
-        plugin_title: "Boom Table",
-        row_col_wrapper: "_",
-    },
-    plugin_id: plugin_id,
-    valueNameOptions: [{
-        text: "Min",
-        value: "min",
-    },
-    {
-        text: "Max",
-        value: "max"
-    },
-    {
-        text: "Average",
-        value: "avg"
-    },
-    {
-        text: "Current",
-        value: "current"
-    },
-    {
-        text: "Total",
-        value: "total"
-    }
-    ]
-};
 
 export {
-    config,
-    kbn,
-    loadPluginCss,
-    MetricsPanelCtrl,
-    TimeSeries,
-    utils,
+    defaultPattern
 };
